Working and Fast (Monte Carlo Method)
This uses a basic Monte Carlo method to make an optimized decision tree. Being that it a statistical approximation, it doesn't guarantee the exact answer each time, but it doesn't require too many pass-throughs to get the correct answer. 1000 trials gave me 1074 each time with a run time of about 0.04 seconds.